add error handling to controllers

+++ backend/src/controllers/epic.js
@@ -1,6 +1,17 @@
+import errorCodeToMessageMap from "../constants/errorMessages.js";
 import epicService from "../services/epicService.js";
 import projectService from "../services/projectService.js";
-import { NO_PROJECT_ERROR } from "./project.js";
+import ticketService from "../services/ticketService.js";
+import ErrorResponse from "../utils/ErrorResponse.js";
+import { isProjectDeleted } from "../utils/projectUtils.js";
+import { PROJECT_NOT_FOUND_ERROR } from "./project.js";
+import { TICKET_NOT_FOUND_ERROR } from "./ticket.js";
+
+const EPIC_NOT_FOUND_ERROR = new ErrorResponse(
+  errorCodeToMessageMap["EPIC_NOT_FOUND"],
+  "EPIC_NOT_FOUND",
+  404,
+);
 
 export const createEpic = async (req, res, next) => {
   const { title, projectId } = req.body;
@@ -8,12 +19,12 @@ export const createEpic = async (req, res, next) => {
   try {
     const project = await projectService.getProject(projectId);
 
-    if (project == null) {
-      throw NO_PROJECT_ERROR;
+    if (isProjectDeleted(project)) {
+      throw PROJECT_NOT_FOUND_ERROR;
     }
 
     const epic = await epicService.createEpic(projectId, { title });
-    res.status(201).send(epic);
+    res.status(201).send({ data: epic });
   } catch (error) {
     next(error);
   }
@@ -24,6 +35,12 @@ export const updateEpic = async (req, res, next) => {
   const { title, description } = req.body;
 
   try {
+    const [epic] = await epicService.getEpics([epicId]);
+
+    if (epic == null) {
+      throw EPIC_NOT_FOUND_ERROR;
+    }
+
     await epicService.updateEpic(epicId, { title, description });
     res.sendStatus(204);
   } catch (error) {
@@ -36,7 +53,7 @@ export const getEpics = async (req, res, next) => {
 
   try {
     const epics = await epicService.getEpics(epicIds);
-    res.send(epics);
+    res.send({ data: epics });
   } catch (error) {
     next(error);
   }
@@ -47,6 +64,17 @@ export const addTicketToEpic = async (req, res, next) => {
   const { ticketId } = req.body;
 
   try {
+    const [epic] = await epicService.getEpics([epicId]);
+    const [ticket] = await ticketService.getTickets([ticketId]);
+
+    if (epic == null) {
+      throw EPIC_NOT_FOUND_ERROR;
+    }
+
+    if (ticket == null) {
+      throw TICKET_NOT_FOUND_ERROR;
+    }
+
     await epicService.addTicketToEpic(ticketId, epicId);
     res.sendStatus(204);
   } catch (error) {
@@ -59,6 +87,17 @@ export const removeTicketFromEpic = async (req, res, next) => {
   const { ticketId } = req.body;
 
   try {
+    const [epic] = await epicService.getEpics([epicId]);
+    const [ticket] = await ticketService.getTickets([ticketId]);
+
+    if (epic == null) {
+      throw EPIC_NOT_FOUND_ERROR;
+    }
+
+    if (ticket == null) {
+      throw TICKET_NOT_FOUND_ERROR;
+    }
+
     await epicService.removeTicketFromEpic(ticketId, epicId);
     res.sendStatus(204);
   } catch (error) {
@@ -70,6 +109,12 @@ export const deleteEpic = async (req, res, next) => {
   const { epicId } = req.params;
 
   try {
+    const [epic] = await epicService.getEpics([epicId]);
+
+    if (epic == null) {
+      throw EPIC_NOT_FOUND_ERROR;
+    }
+
     await epicService.deleteEpic(epicId);
     res.sendStatus(204);
   } catch (error) {
