+++ backend/src/controllers/project.js
@@ -1,8 +1,10 @@
 import errorCodeToMessageMap from "../constants/errorMessages.js";
 import projectService from "../services/projectService.js";
 import ErrorResponse from "../utils/ErrorResponse.js";
+import objectUtils from "../utils/objectUtils.js";
+import { isProjectDeleted } from "../utils/projectUtils.js";
 
-export const NO_PROJECT_ERROR = new ErrorResponse(
+export const PROJECT_NOT_FOUND_ERROR = new ErrorResponse(
   errorCodeToMessageMap["PROJECT_NOT_FOUND"],
   "PROJECT_NOT_FOUND",
   404,
@@ -13,7 +15,7 @@ export const createProject = async (req, res, next) => {
 
   try {
     const newProject = await projectService.createProject({ title, adminId: [adminId], type });
-    res.status(201).send(newProject);
+    res.status(201).send({ data: newProject });
   } catch (error) {
     next(error);
   }
@@ -24,11 +26,11 @@ export const getProject = async (req, res, next) => {
 
   try {
     const project = await projectService.getProject(projectId);
-    if (project == null || project.isDeleted) {
-      throw NO_PROJECT_ERROR;
+    if (isProjectDeleted(project)) {
+      throw PROJECT_NOT_FOUND_ERROR;
     }
 
-    res.send(project);
+    res.send({ data: project });
   } catch (error) {
     next(error);
   }
@@ -41,11 +43,9 @@ export const addMemberToProject = async (req, res, next) => {
   try {
     const project = await projectService.getProject(projectId);
 
-    if (project == null || project.isDeleted) {
-      throw NO_PROJECT_ERROR;
+    if (isProjectDeleted(project)) {
+      throw PROJECT_NOT_FOUND_ERROR;
     }
-
-    // TODO: verify if userId is a valid user
 
     await projectService.addMembersToProject([userId], projectId);
     res.sendStatus(204);
@@ -61,11 +61,9 @@ export const addMembersToProject = async (req, res, next) => {
   try {
     const project = await projectService.getProject(projectId);
 
-    if (project == null || project.isDeleted) {
-      throw NO_PROJECT_ERROR;
+    if (isProjectDeleted(project)) {
+      throw PROJECT_NOT_FOUND_ERROR;
     }
-
-    // TODO: verify if userIds is a valid user
 
     await projectService.addMembersToProject(userIds, projectId);
     res.sendStatus(204);
@@ -81,11 +79,10 @@ export const removeMemberFromProject = async (req, res, next) => {
   try {
     const project = await projectService.getProject(projectId);
 
-    if (project == null || project.isDeleted) {
-      throw NO_PROJECT_ERROR;
+    if (isProjectDeleted(project)) {
+      throw PROJECT_NOT_FOUND_ERROR;
     }
 
-    // TODO: verify if userIds is a valid user
     await projectService.removeMemberFromProject(userId, projectId);
     res.sendStatus(204);
   } catch (error) {
@@ -99,17 +96,20 @@ export const updateProject = async (req, res, next) => {
 
   try {
     const project = await projectService.getProject(projectId);
-    if (project == null || project.isDeleted) {
-      throw NO_PROJECT_ERROR;
+
+    if (isProjectDeleted(project)) {
+      throw PROJECT_NOT_FOUND_ERROR;
     }
 
-    await projectService.updateProject(projectId, {
+    const definedKeys = objectUtils.removeUndefinedKeysFromObject({
       title,
       description,
       announcement,
       picture,
       type,
     });
+
+    await projectService.updateProject(projectId, definedKeys);
     res.sendStatus(204);
   } catch (error) {
     next(error);
@@ -121,7 +121,7 @@ export const getProjectsOfUser = async (req, res, next) => {
 
   try {
     const projects = await projectService.getProjectsOfUser(userId);
-    res.send(projects);
+    res.send({ data: projects });
   } catch (error) {
     next(error);
   }
@@ -134,8 +134,8 @@ export const promoteMemberToAdministrator = async (req, res, next) => {
   try {
     const project = await projectService.getProject(projectId);
 
-    if (project == null || project.isDeleted) {
-      throw NO_PROJECT_ERROR;
+    if (isProjectDeleted(project)) {
+      throw PROJECT_NOT_FOUND_ERROR;
     }
 
     await projectService.promoteMemberToAdmin(userId, projectId);
@@ -152,8 +152,8 @@ export const demoteAdmintoMember = async (req, res, next) => {
   try {
     const project = await projectService.getProject(projectId);
 
-    if (project == null || project.isDeleted) {
-      throw NO_PROJECT_ERROR;
+    if (isProjectDeleted(project)) {
+      throw PROJECT_NOT_FOUND_ERROR;
     }
 
     if (project.adminIds.length <= 1) {
@@ -175,6 +175,12 @@ export const deleteProject = async (req, res, next) => {
   const { projectId } = req.params;
 
   try {
+    const project = await projectService.getProject(projectId);
+
+    if (isProjectDeleted(project)) {
+      throw PROJECT_NOT_FOUND_ERROR;
+    }
+
     await projectService.deleteProject(projectId);
     res.sendStatus(204);
   } catch (error) {
