+++ backend/src/controllers/ticket.js
@@ -1,5 +1,15 @@
+import errorCodeToMessageMap from "../constants/errorMessages.js";
 import projectService from "../services/projectService.js";
 import ticketService from "../services/ticketService.js";
+import ErrorResponse from "../utils/ErrorResponse.js";
+import { isProjectDeleted } from "../utils/projectUtils.js";
+import { PROJECT_NOT_FOUND_ERROR } from "./project.js";
+
+export const TICKET_NOT_FOUND_ERROR = new ErrorResponse(
+  errorCodeToMessageMap["TICKET_NOT_FOUND"],
+  "TICKET_NOT_FOUND",
+  404,
+);
 
 export const createTicket = async (req, res, next) => {
   const { title, projectId, priority, type } = req.body;
@@ -7,12 +17,12 @@ export const createTicket = async (req, res, next) => {
   try {
     const project = await projectService.getProject(projectId);
 
-    if (project == null || project.isDeleted) {
-      throw NO_PROJECT_ERROR;
+    if (isProjectDeleted(project)) {
+      throw PROJECT_NOT_FOUND_ERROR;
     }
 
     const newTicket = ticketService.createTicket(projectId, { title, priority, type });
-    res.status(201).send(newTicket);
+    res.status(201).send({ data: newTicket });
   } catch (error) {
     next(error);
   }
@@ -33,6 +43,12 @@ export const updateTicket = async (req, res, next) => {
   } = req.body;
 
   try {
+    const [ticket] = await ticketService.getTickets([ticketId]);
+
+    if (ticket == null) {
+      throw TICKET_NOT_FOUND_ERROR;
+    }
+
     await ticketService.updateTicket(ticketId, {
       title,
       description,
@@ -44,6 +60,7 @@ export const updateTicket = async (req, res, next) => {
       assigneeId,
       storyPoints,
     });
+
     res.sendStatus(204);
   } catch (error) {
     next(error);
@@ -52,9 +69,14 @@ export const updateTicket = async (req, res, next) => {
 
 export const deleteTicket = async (req, res, next) => {
   const { ticketId } = req.params;
-  // const { projectId } = req.body;
 
   try {
+    const [ticket] = await ticketService.getTickets([ticketId]);
+
+    if (ticket == null) {
+      throw TICKET_NOT_FOUND_ERROR;
+    }
+
     await ticketService.deleteTicket(ticketId);
     res.sendStatus(204);
   } catch (error) {
@@ -67,7 +89,7 @@ export const getTickets = async (req, res, next) => {
 
   try {
     const tickets = await ticketService.getTickets(ticketIds);
-    res.send(tickets);
+    res.send({ data: tickets });
   } catch (error) {
     next(error);
   }
@@ -78,7 +100,12 @@ export const getTicket = async (req, res, next) => {
 
   try {
     const [ticket] = await ticketService.getTickets([ticketId]);
-    res.send(ticket);
+
+    if (ticket == null) {
+      throw TICKET_NOT_FOUND_ERROR;
+    }
+
+    res.send({ data: ticket });
   } catch (error) {
     next(error);
   }
